ml-server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
langchain==0.3.16
langchain-community==0.3.16
langchain-openai==0.2.14
langchain-qdrant==0.2.1
beautifulsoup4==4.12.3
pydantic==2.10.5
mcp[cli]
httpx
openai-agents
mem0ai